fix: Don't check for padding when length is 0
ANSISegment:
fix: Previously, when the segment didn't hold any actual data, the
toString() function would crash trying to construct an empty string.
This was because the string is constructed from the beginning of the
data vector minus two bytes (for the header byte and length byte) to
the end of the data vector, minus a possible pad byte. When the data
length byte was zero, the check for the padding interpreted the last
byte as the pad byte, when this was the length byte, and thus an invalid
string would be constructed. Now, this case is checked at the beginning.